+++ app/controllers/tell_us_more_about_the_new_job_controller.rb
@@ -1,0 +1,5 @@
+class TellUsMoreAboutTheNewJobController < FormsController
+  def self.show_rule_sets(change_report)
+    super << change_report.change_type_new_job?
+  end
+end
