+++ app/views/tell_us_about_the_new_job/edit.html.erb
@@ -3,19 +3,19 @@
 <% content_for :card_body do %>
   <%= fields_for(:form, @form, builder: CfaFormBuilder) do |f| %>
     <%= f.cfa_input_field :company_name, "What is the name of the company?" %>
+
+    <div class="notice">
+      <p>The following questions are optional, but they'll help the county verify your change. Try to answer them.</p>
+    </div>
+
     <%= f.cfa_input_field :manager_name,
                           t(self_or_other_member_translation_key(".manager_name"), name: current_change_report.member.first_name),
-                          help_text: t(self_or_other_member_translation_key(".manager_name_help")) %>
-    <%= f.cfa_input_field :manager_phone_number, t(self_or_other_member_translation_key(".manager_phone_number")), classes: ["form-width--phone"] %>
-    <%= f.cfa_date_select :first_day,
-                          t(self_or_other_member_translation_key(".first_day")),
-                          help_text: "Estimate if you're not sure.",
-                          options: { start_year: Time.now.year, end_year: Time.now.year - 3 } %>
-    <%= f.cfa_radio_set :paid_yet,
-                        label_text: t(self_or_other_member_translation_key(".paid_yet")),
-                        collection: [
-                            { value: :yes, label: "Yes" },
-                            { value: :no, label: "No" },
-                        ] %>
+                          help_text: t(self_or_other_member_translation_key(".manager_name_help")),
+                          optional: true %>
+    <%= f.cfa_input_field :manager_phone_number,
+                          t(self_or_other_member_translation_key(".manager_phone_number")),
+                          classes: ["form-width--phone"],
+                          help_text: "You can provide a general phone number if you don’t know a direct line.",
+                          optional: true %>
   <% end %>
 <% end %>
