+++ app/views/tell_us_more_about_the_new_job/edit.html.erb
@@ -1,0 +1,16 @@
+<% content_for :card_title, "Tell us more about the new job." %>
+
+<% content_for :card_body do %>
+  <%= fields_for(:form, @form, builder: CfaFormBuilder) do |f| %>
+    <%= f.cfa_date_select :first_day,
+      t(self_or_other_member_translation_key(".first_day")),
+      help_text: "Estimate if you're not sure.",
+      options: { start_year: Time.now.year, end_year: Time.now.year - 3 } %>
+    <%= f.cfa_radio_set :paid_yet,
+      label_text: t(self_or_other_member_translation_key(".paid_yet")),
+      collection: [
+        { value: :yes, label: "Yes" },
+        { value: :no, label: "No" },
+      ] %>
+  <% end %>
+<% end %>
