+++ spec/features/report_new_job_spec.rb
@@ -53,6 +53,9 @@ RSpec.feature "Reporting a change", :a11y, :js do
       fill_in "What is the name of the company?", with: "Abc Corp"
       fill_in "What is the name of Jane's supervisor or manager?", with: "My boss"
       fill_in "What is this person's phone number?", with: "999-888-7777"
+      proceed_with "Continue"
+
+      expect(page.title).to eq "Tell us more about the new job. | ReportChangesColorado.org"
       select "February", from: "form[first_day_month]"
       select "2", from: "form[first_day_day]"
       select "2018", from: "form[first_day_year]"
